Move ts/Vec to ts/Map
