+++ lib/ts/test_Vec.cc
@@ -25,7 +25,182 @@
 #include <cstdint>
 #include <cstdio>
 #include "ts/ink_assert.h"
-#include "ts/Vec.h"
+#include "ts/Map.h"
+
+// Intervals store sets in interval format (e.g. [1..10][12..12]).
+// Inclusion test is by binary search on intervals.
+// Deletion is not supported
+class Intervals : public Vec<int>
+{
+public:
+  void insert(int n);
+  bool in(int n) const;
+};
+
+// UnionFind supports fast unify and finding of
+// 'representitive elements'.
+// Elements are numbered from 0 to N-1.
+class UnionFind : public Vec<int>
+{
+public:
+  // set number of elements, initialized to singletons, may be called repeatedly to increase size
+  void size(int n);
+  // return representitive element
+  int find(int n);
+  // unify the sets containing the two elements
+  void unify(int n, int m);
+};
+
+// binary search over intervals
+static int
+i_find(const Intervals *i, int x)
+{
+  ink_assert(i->n);
+  int l = 0, h = i->n;
+Lrecurse:
+  if (h <= l + 2) {
+    if (h <= l) {
+      return -(l + 1);
+    }
+    if (x < i->v[l] || x > i->v[l + 1]) {
+      return -(l + 1);
+    }
+    return h;
+  }
+  int m = (((h - l) / 4) * 2) + l;
+  if (x > i->v[m + 1]) {
+    l = m;
+    goto Lrecurse;
+  }
+  if (x < i->v[m]) {
+    h = m;
+    goto Lrecurse;
+  }
+  return (l + 1);
+}
+
+bool
+Intervals::in(int x) const
+{
+  if (!n) {
+    return false;
+  }
+  if (i_find(this, x) > 0) {
+    return true;
+  }
+  return false;
+}
+
+// insert into interval with merge
+void
+Intervals::insert(int x)
+{
+  if (!n) {
+    add(x);
+    add(x);
+    return;
+  }
+  int l = i_find(this, x);
+  if (l > 0) {
+    return;
+  }
+  l = -l - 1;
+
+  if (x > v[l + 1]) {
+    if (x == v[l + 1] + 1) {
+      v[l + 1]++;
+      goto Lmerge;
+    }
+    l += 2;
+    if (l < (int)n) {
+      if (x == v[l] - 1) {
+        v[l]--;
+        goto Lmerge;
+      }
+    }
+    goto Lmore;
+  } else {
+    ink_assert(x < v[l]);
+    if (x == v[l] - 1) {
+      v[l]--;
+      goto Lmerge;
+    }
+    if (!l) {
+      goto Lmore;
+    }
+    l -= 2;
+    if (x == v[l + 1] + 1) {
+      v[l + 1]++;
+      goto Lmerge;
+    }
+  }
+Lmore:
+  fill(n + 2);
+  if (n - 2 - l > 0) {
+    memmove(v + l + 2, v + l, sizeof(int) * (n - 2 - l));
+  }
+  v[l]     = x;
+  v[l + 1] = x;
+  return;
+Lmerge:
+  if (l) {
+    if (v[l] - v[l - 1] < 2) {
+      l -= 2;
+      goto Ldomerge;
+    }
+  }
+  if (l < (int)(n - 2)) {
+    if (v[l + 2] - v[l + 1] < 2) {
+      goto Ldomerge;
+    }
+  }
+  return;
+Ldomerge:
+  memmove(v + l + 1, v + l + 3, sizeof(int) * (n - 3 - l));
+  n -= 2;
+  goto Lmerge;
+}
+
+void
+UnionFind::size(int s)
+{
+  size_t nn = n;
+  fill(s);
+  for (size_t i = nn; i < n; i++) {
+    v[i] = -1;
+  }
+}
+
+int
+UnionFind::find(int n)
+{
+  int i, t;
+  for (i = n; v[i] >= 0; i = v[i]) {
+    ;
+  }
+  while (v[n] >= 0) {
+    t    = n;
+    n    = v[n];
+    v[t] = i;
+  }
+  return i;
+}
+
+void
+UnionFind::unify(int n, int m)
+{
+  n = find(n);
+  m = find(m);
+  if (n != m) {
+    if (v[m] < v[n]) {
+      v[m] += (v[n] - 1);
+      v[n] = m;
+    } else {
+      v[n] += (v[m] - 1);
+      v[m] = n;
+    }
+  }
+}
 
 static void
 test_append()
